Reworked File Structure
