+++ src/__init__.py
@@ -1,6 +1,5 @@
 #Initiallises everything.
 
 import src.game
-import src.main
 import src.settings
 import src.utils
